+++ bossdata/plate.py
@@ -126,11 +126,17 @@ class Plan(object):
                              .format(fiber, self.num_fibers, self.plate))
         return 1 if fiber <= self.num_fibers // 2 else 2
 
-    def get_exposure_name(self, sequence_number, camera, fiber, calibrated=True):
+    def get_exposure_name(self, sequence_number, camera, fiber, ftype='spCFrame'):
         """Get the name of a science exposure from this plan.
 
-        Use the exposure name to locate the calibrated spCFrame-{EXPNAME}.fits and
-        uncalibrated spFrame-{EXPNAME}.fits.gz files for individual exposures.
+        Use the exposure name to locate FITS data files associated with
+        individual exposures.  The supported file types are:
+        :datamodel:`spCFrame <PLATE4/spCFrame.html>`,
+        :datamodel:`spFrame <PLATE4/spFrame.html>`,
+        :datamodel:`spFluxcalib <PLATE4/spFluxcalib.html>` and
+        :datamodel:`spFluxcorr <PLATE4/spFluxcorr.html>`.
+        Note that this method returns None when the requested exposure is not present
+        in the plan, so the return value should always be checked.
 
         Args:
             sequence_number(int): Science exposure sequence number, counting from zero.
@@ -138,16 +144,16 @@ class Plan(object):
             fiber(int): Fiber number to identify which spectrograph to use, which must
                 be in the range 1-1000 (or 1-640 for plate < 3510).
             camera(str): Must be 'blue' or 'red'.
-            calibrated(bool): Returns the name of the calibrated (spCFrame) file rather
-                than the un-calibrated (spFrame) file.
+            ftype(str): Type of exposure file whose name to return.  Must be one of
+                spCFrame, spFrame, spFluxcalib, spFluxcorr.  An spCFrame is assumed
+                to be uncompressed, and all other files are assumed to be compressed.
 
         Returns:
-            str: Exposure name of the form [prefix]-[cc]-[eeeeeeee].[ext] where [cc]
+            str: Exposure name of the form [ftype]-[cc]-[eeeeeeee].[ext] where [cc]
                 identifies the spectrograph (one of b1,r1,b2,r2) and [eeeeeeee] is the
-                zero-padded exposure number. For calibrated exposures, [prefix] is
-                "spCFrame" and [ext] is "fits".  For un-calibrated exposures, [prefix]
-                is "spFrame" and [ext] is "fits.gz". Returns None if the name is
-                unknown for this camera and fiber combination.
+                zero-padded exposure number. The extension [ext] is "fits" for
+                spCFrame files and "fits.gz" for all other file types. Returns None if
+                the name is unknown for this camera and fiber combination.
 
         Raises:
             ValueError: one of the inputs is invalid.
@@ -160,18 +166,19 @@ class Plan(object):
                 fiber, self.num_fibers, self.plate))
         if camera not in ('blue', 'red'):
             raise ValueError('Invalid camera ({}) must be blue or red.'.format(camera))
+        if ftype not in ('spCFrame', 'spFrame', 'spFluxcalib', 'spFluxcorr'):
+            raise ValueError('Invalid file type ({}) must be one of: '.format(ftype) +
+                             'spCFrame, spFrame, spFluxcalib, spFluxcorr.')
 
         spectrograph = camera[0] + str(self.get_spectrograph_index(fiber))
         exposure_info = self.exposures['science'][sequence_number]
         if spectrograph not in exposure_info['SPECS']:
             return None
         exposure_id = exposure_info['EXPID']
-        if calibrated:
-            prefix, ext = 'spCFrame', 'fits'
-        else:
-            prefix, ext = 'spFrame', 'fits.gz'
-
-        return '{0}-{1}-{2:08d}.{3}'.format(prefix, spectrograph, exposure_id, ext)
+        name = '{0}-{1}-{2:08d}.fits'.format(ftype, spectrograph, exposure_id)
+        if ftype != 'spCFrame':
+            name += '.gz'
+        return name
 
 
 class TraceSet(object):